1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,11 +11,11 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
-	if ((int)size < 0)
+	if (size < 0)
 		free(array);
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
 		action(array[i]);
 	}
